BIGINT switch

+++ util/execution/table_generator/table_generator.cpp
@@ -317,7 +317,7 @@ void TableGenerator::GenerateMiniRunnerIndexes() {
   std::vector<TableInsertMeta> table_metas;
   std::vector<uint32_t> idx_key = {1, 2, 4, 8, 15};
   std::vector<uint32_t> row_nums = {1, 10, 100, 500, 1000, 2000, 5000, 10000, 20000, 50000, 100000, 500000, 1000000};
-  std::vector<type::TypeId> types = {type::TypeId::INTEGER,type::TypeId::DECIMAL};
+  std::vector<type::TypeId> types = {type::TypeId::INTEGER,type::TypeId::BIGINT};
   for (auto row_num : row_nums) {
     for (type::TypeId type : types) {
       auto table_name = GenerateTableIndexName(type, row_num);
@@ -414,7 +414,7 @@ std::vector<TableGenerator::TableInsertMeta> TableGenerator::GenerateMiniRunnerT
   std::vector<TableInsertMeta> table_metas;
   std::vector<uint32_t> row_nums = {1,    3,    5,     7,     10,    50,     100,    500,    1000,
                                     2000, 5000, 10000, 20000, 50000, 100000, 200000, 500000, 1000000};
-  std::vector<type::TypeId> types = {type::TypeId::INTEGER, type::TypeId::DECIMAL};
+  std::vector<type::TypeId> types = {type::TypeId::INTEGER, type::TypeId::BIGINT};
   for (int col_num = 15; col_num <= 15; col_num++) {
     for (uint32_t row_num : row_nums) {
       // Cardinality of the last column
